\name{MRAID-package}
\alias{MRAID-package}
\alias{MRAID}
\docType{package}
\title{\packageTitle{MRAID}}
\description{MRAID(Mendelian Randomization with Automated Instrument Determination)}
\details{
 \tabular{ll}{
Package: \tab MRAID\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2021-06-05\cr
License: \tab GPL-3\cr
}
}
\author{
   Zhongshang Yuan and Lu Liu
  
  Maintainer: Zhongshang Yuan <yuanzhongshang@sdu.edu.cn>
}

% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
