% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRAID.R
\name{MRAID}
\alias{MRAID}
\title{The function of MRAID method two sample Mendelian Randomization}
\usage{
MRAID(
  Zscore_1,
  Zscore_2,
  Sigma1sin,
  Sigma2sin,
  samplen1,
  samplen2,
  Gibbsnumber = 1000,
  burninproportion = 0.2,
  pi_beta_shape = 0.5,
  pi_beta_scale = 4.5,
  pi_c_shape = 0.5,
  pi_c_scale = 9.5,
  pi_1_shape = 0.5,
  pi_1_scale = 1.5,
  pi_0_shape = 0.05,
  pi_0_scale = 9.95
)
}
\arguments{
\item{Zscore_1}{the Zscore vector of the SNP effect size vector for the exposure}

\item{Zscore_2}{the Zscore vector of the SNP effect size vector for the outcome}

\item{Sigma1sin}{the LD matrix for the SNPs in the exposure GWAS data}

\item{Sigma2sin}{the LD matrix for the SNPs in the outcome GWAS data,both Sigma2sin and sigma1sin are often from the same reference panel}

\item{samplen1}{the sample size of exposure GWAS}

\item{samplen2}{the sample size of outcome GWAS}

\item{Gibbsnumber}{the number of Gibbs sampling iterations with the default to be 1000}

\item{burninproportion}{the proportion to burn in from Gibbs sampling iterations, with default to be 20%}

\item{pi_beta_shape}{the prior shape paramter for pi_beta with the default to be 0.5}

\item{pi_beta_scale}{the prior scale paramter for pi_beta with the default to be 4.5}

\item{pi_c_shape}{the prior shape paramter for pi_c with the default to be 0.5}

\item{pi_c_scale}{the prior shape paramter for pi_c with the default to be 9.5}

\item{pi_1_shape}{the prior shape paramter for pi_1 with the default to be 0.5}

\item{pi_1_scale}{the prior scale paramter for pi_1 with the default to be 1.5}

\item{pi_0_shape}{the prior shape paramter for pi_0 with the default to be 0.05}

\item{pi_0_scale}{the prior scale paramter for pi_0 with the default to be 9.95}
}
\value{
A list of estimated parameters including the p values for the causal effect test 
\item{causal_effect}{The estimate of causal effect}
\item{causal_pvalue}{The p value for the causal effect}
\item{correlated_pleiotropy_effect}{The estimate of correlated pleiotropy effect}
\item{sigmaeta}{The variance estimate for the uncorrelated pleiotropy effect}
\item{sigmabeta}{The variance estimate for the SNP effect sizes on the exposure}
\item{sigma_error_1}{The variance estimate of the error in exposure GWAS model}
\item{sigma_error_2}{The variance estimate of the error in outcome GWAS model}
}
\description{
MRAID is able to model an initial set of candidate SNP instruments that are in high LD with each other and perform automated instrument selection to identify suitable SNPs to serve as instrumental variables. MRAID simultaneously accounts for both uncorrelated and correlated horizontal pleiotropy, relies on a scalable sampling-based inference algorithm to perform numerical integration, circumventing the difficulty in likelihood function
}
