% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MRAID_CPP}
\alias{MRAID_CPP}
\title{MRAID}
\usage{
MRAID_CPP(
  betaxin,
  betayin,
  Sigma1sin,
  Sigma2sin,
  samplen1,
  samplen2,
  Gibbsnumberin,
  burninproportion,
  initial_betain,
  pi_beta_shape_in,
  pi_beta_scale_in,
  pi_c_shape_in,
  pi_c_scale_in,
  pi_1_shape_in,
  pi_1_scale_in,
  pi_0_shape_in,
  pi_0_scale_in，
  maxvarin，
  alphain
)
}
\description{
MRAID
}
